Avoid `create` in `Dag`
